add chunking + restructuring
