+++ pyproject.toml
@@ -4,14 +4,16 @@ version = "0.1.0"
 description = ""
 authors = ["Your Name <you@example.com>"]
 
+include = ["README.md", "setup.cfg", "src/**/*.py"]
+
 [tool.poetry.dependencies]
 python = "^3.7"
 chardet = "^3.0.4"
 click = "^7.1.2"
 sepaxml = "^2.3.0"
 unidecode = "^1.1.1"
-setuptools = "^50.3.0"
 schwifty = "^2020.9.0"
+tqdm = "^4.50.1"
 
 [tool.poetry.dev-dependencies]
 flake8 = "^3.8.3"
@@ -24,6 +26,9 @@ flake8-quotes = "^3.2.0"
 pep8-naming = "^0.11.1"
 mypy = "^0.782"
 pydocstyle = "^5.0.2"
+
+[tool.poetry.scripts]
+xmler = 'xmler.__main__:cli'
 
 [build-system]
 requires = ["poetry>=0.12"]
